docs(gatsby/theme): refactor component 'PatternsStatus' & rename it as 'PatternsStatusTable'
[skip ci]
