--- packages/gatsby-theme-styleguide/src/components/PatternsStatus/__tests__/patternStatus.basic.mock.js
@@ -1,72 +1,0 @@
-export default {
-  Foundations: {
-    totalCount: 2,
-    edges: [
-      {
-        node: {
-          fields: {
-            slug: '/Foundations/Typography/BodyStyles/',
-          },
-          frontmatter: {
-            title: 'Body styles',
-            order: 4,
-            status: {
-              sketch: 'bêta',
-              scss: 'bêta',
-            },
-          },
-        },
-      },
-      {
-        node: {
-          fields: {
-            slug: '/Foundations/BorderRadius/',
-          },
-          frontmatter: {
-            title: 'Border radius',
-            order: 8,
-            status: {
-              sketch: 'bêta',
-              scss: 'bêta',
-            },
-          },
-        },
-      },
-    ],
-  },
-  Components: {
-    totalCount: 2,
-    edges: [
-      {
-        node: {
-          fields: {
-            slug: '/Components/Typography/BodyStyles/',
-          },
-          frontmatter: {
-            title: 'Body styles',
-            order: 4,
-            status: {
-              sketch: 'bêta',
-              scss: 'bêta',
-            },
-          },
-        },
-      },
-      {
-        node: {
-          fields: {
-            slug: '/Components/Typography/BodyStyles/',
-          },
-          frontmatter: {
-            title: 'Border radius',
-            order: 8,
-            status: {
-              sketch: 'bêta',
-              scss: 'bêta',
-            },
-          },
-        },
-      },
-    ],
-  },
-}
