--- packages/gatsby-theme-styleguide/src/components/PatternsStatus/index.js
@@ -1,1 +1,0 @@
-export { default } from './PatternsStatus'
