+++ packages/gatsby-theme-styleguide/src/components/PatternsStatusTable/index.js
@@ -1,0 +1,1 @@
+export { default } from './PatternsStatusTable'
