+++ packages/gatsby-theme-styleguide/src/components/PatternsStatusTable/patternstatus.module.css
@@ -1,0 +1,65 @@
+/* PatternsStatus */
+
+/** StatusTable */
+.statusTable {
+  font-size: 0.875rem;
+  color: #525252;
+}
+
+.statusTable td,
+.statusTable th {
+  padding: 0.8rem 0.4rem 1rem;
+  line-height: 1.3;
+}
+
+.statusTable tr {
+  border: 0 !important;
+}
+
+.statusTable th {
+  border: 0 !important;
+}
+
+/** StatusMeaning */
+.statusMeaning {
+  display: flex;
+  align-items: center;
+}
+
+/** Discipline */
+.discipline {
+  border-bottom: solid 1px #554f52;
+  font-size: 1rem;
+  line-height: 2.5em;
+}
+
+/** PatternName */
+.patternName {
+  display: block;
+  font-weight: bold;
+  padding-left: 0.5rem;
+  width: 110px;
+  max-width: 110px;
+  color: #525252;
+}
+
+.patternName:hover {
+  text-decoration: underline;
+}
+
+/** Row */
+.row {
+  border: 0;
+}
+
+.row:nth-child(odd) {
+  background-color: #f5f5f5;
+}
+
+/** Platform */
+.platform {
+  font-weight: normal;
+  color: #999;
+  padding-top: 0.6rem !important;
+  padding-bottom: 0.4rem !important;
+}
